Moving monitoring context manager to the monitoring module for reuse by other entrypoints
