+++ src/clusterfuzz/_internal/metrics/monitor.py
@@ -18,9 +18,11 @@
 
 import bisect
 import collections
+import contextlib
 import functools
 import itertools
 import re
+import signal
 import threading
 import time
 from typing import List
@@ -130,6 +132,23 @@ def _flush_metrics():
       logs.warning(f'Failed to flush metrics: {e}')
     else:
       logs.error(f'Failed to flush metrics: {e}')
+
+
+def handle_sigterm(signo, stack_frame):  #pylint: disable=unused-argument
+  logs.info('Handling sigterm, stopping monitoring daemon.')
+  stop()
+  logs.info('Sigterm handled, metrics flushed.')
+
+
+@contextlib.contextmanager
+def wrap_with_monitoring():
+  """Wraps execution so we flush metrics on exit"""
+  try:
+    initialize()
+    signal.signal(signal.SIGTERM, handle_sigterm)
+    yield
+  finally:
+    stop()
 
 
 class _MonitoringDaemon():
